feat: migrate utils to multiplatform
