+++ foundation/utils/src/commonMain/kotlin/io/github/alaksion/invoicer/foundation/utils/logger/InvoicerLogger.kt
@@ -1,0 +1,6 @@
+package io.github.alaksion.invoicer.foundation.utils.logger
+
+interface InvoicerLogger {
+    fun logDebug(message: String, key: String)
+    fun logError(message: String, key: String, throwable: Throwable?)
+}
