+++ foundation/utils/src/commonMain/kotlin/io/github/alaksion/invoicer/foundation/utils/logger/MutedLogger.kt
@@ -1,0 +1,6 @@
+package io.github.alaksion.invoicer.foundation.utils.logger
+
+object MutedLogger : InvoicerLogger {
+    override fun logDebug(message: String, key: String) = Unit
+    override fun logError(message: String, key: String, throwable: Throwable?) = Unit
+}
