+++ foundation/utils/src/commonMain/kotlin/io/github/alaksion/invoicer/foundation/utils/money/MoneyFormat.kt
