+++ foundation/utils/src/iosMain/kotlin/io/github/alaksion/invoicer/foundation/utils/logger/IosInvoicerLogger.kt
@@ -1,0 +1,8 @@
+package io.github.alaksion.invoicer.foundation.utils.logger
+
+internal class IosInvoicerLogger : InvoicerLogger {
+
+    override fun logDebug(message: String, key: String) = Unit
+
+    override fun logError(message: String, key: String, throwable: Throwable?) = Unit
+}
